requirements.txt
pandas==2.2.3
matplotlib==3.9.2
python-dotenv==1.1.0
openai
streamlit
